Change to dbus autostarting the daemon.
Previously the daemon needed to be started manually or
via autostart. Now the daemon is started by a dbus call
to the daemon. This allowed the funky code which created
a main window and showing it prior to deleting a system
tray icon and hiding it immediately thereafter could be
removed since if the daemon goes away, now it will just
be restarted if another window is docked.