Use dbus autostart feature to start the daemon
Changed so the daemon is autostarted by the dbus
service rather than starting manually or by a
session autostart. This also allowed the removal
of the kludgy code where a main window was created
and only "shown" when the system tray icon was
deleted and the "hidden" again, to keep the daemon
from ending execution when the last window was
undocked. Now, it just starts again when the
next window is docked.